// src/core/packages/saved-objects/migration-server-internal/src/zdt/state/types.d.ts
export interface BaseState extends ControlState {
    readonly retryCount: number;
    readonly skipRetryReset: boolean;
    readonly retryDelay: number;
    readonly logs: MigrationLog[];
    /**
     * When true, will fully skip document migration, and will transition directly to DONE
     * after the INDEX_STATE_UPDATE_DONE stage.
     *
     * This flag is set to `true` in the following scenarios:
     * - on nodes without the `migrator` role, the flag will always be `true`.
     * - if the migrator create the index, the workflow will set the flag to `true` given there is nothing to migrate.
     */
    readonly skipDocumentMigration: boolean;
}
/** Initial state before any action is performed */
export interface InitState extends BaseState {
    readonly controlState: 'INIT';
}
/**
 * Common state properties available after the `INIT` stage
 */
export interface PostInitState extends BaseState {
    /**
     * The index we're currently migrating.
     */
    readonly currentIndex: string;
    /**
     * The aliases that are already present for the current index.
     */
    readonly aliases: string[];
    /**
     * The alias actions to perform to update the aliases.
     */
    readonly aliasActions: AliasAction[];
    /**
     * The *previous* mappings (and _meta), as they were when we resolved the index
     * information. This shouldn't be updated once populated.
     */
    readonly previousMappings: IndexMapping;
    /**
     * The *current* _meta field of the index.
     * All operations updating this field will update in the state accordingly.
     */
    readonly currentIndexMeta: IndexMappingMeta;
    /**
     * The previous algorithm that was last used to migrate this index.
     * Used for v2->zdt state conversion.
     */
    readonly previousAlgorithm: 'zdt' | 'v2';
}
/**
 * Common state properties available after the `DOCUMENTS_UPDATE_INIT` stage
 */
export interface PostDocInitState extends PostInitState {
    readonly excludeOnUpgradeQuery: QueryDslQueryContainer;
    readonly excludeFromUpgradeFilterHooks: Record<string, SavedObjectTypeExcludeFromUpgradeFilterHook>;
    readonly outdatedDocumentsQuery: QueryDslQueryContainer;
    readonly transformRawDocs: TransformRawDocs;
}
export interface OutdatedDocumentsSearchState extends PostDocInitState {
    readonly pitId: string;
    readonly lastHitSortValue: number[] | undefined;
    readonly corruptDocumentIds: string[];
    readonly transformErrors: TransformErrorObjects[];
    readonly progress: Progress;
    readonly hasTransformedDocs: boolean;
}